+++ pymetacode/coding.py
@@ -928,7 +928,12 @@ class FunctionCreator:
     def _function_exists_in_module(self):
         with open(self._module_filename, encoding="utf8") as file:
             contents = file.read()
-        return f"def {self.name}" in contents
+        return any(
+            [
+                name in contents
+                for name in [f"def {self.name}:", f"def {self.name}("]
+            ]
+        )
 
     def _create_function(self):
         context = self.configuration.to_dict()